+++ lootbox.py
@@ -6,13 +6,15 @@ import random
 
 class Lootbox():
     '''Lootbox class'''
-    def __init__(self, lootTable, textColours):
+    def __init__(self, lootTable, inventory, textColours):
         '''
         sets up the class
         lootTable: a dict containing items and their chances
+        inventory: where items are deposited to
         textColours: a dict used to colour text
         '''
         self.lootTable = lootTable
+        self.inventory = inventory
         self.textColours = textColours
         counter = lootTable.copy()
         for item in counter:
@@ -61,9 +63,17 @@ class Lootbox():
         stores the number of items in each rarity
         item: the item you are storing
         '''
+        #to prevent bugs it will create a key in counter if the key
+        #doesn't already exist
         if item in self.counter:
             self.counter[item] += 1
         else:
             self.counter[item] = 1
+        #to prevent bugs it will create a key in inventory if the key
+        #doesn't already exist
+        if item in self.inventory:
+            self.inventory[item] += 1
+        else:
+            self.inventory[item] = 1